Update add book form and add random select page Reorganize API vs App routes
