+++ views/books.html
@@ -4,26 +4,64 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>🦶 Book List</title>
-    <link rel="stylesheet" type="text/css" href="stylesheets/books.css">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
   </head>
-  <body>
-    <h1>Reading List</h1>
-    <div class="reading-list">
-      <div class="bg-transparent card border-0 mx-3 px-1">
-        <h3>Currently Reading</h3>
-        <div class="d-flex align-items-center">
-          <ul class="mb-0 mr-3" id="currently-reading">
-            <li>No books are currently being read.</li>
-          </ul>
+  <body></body>
+  <div class="container mt-5">
+    <div class="row">
+      <div class="col-md-8 offset-md-2">
+        <h1 class="text-center mb-4">Reading List</h1>
+        <div class="card">
+          <div class="card-body">
+            <div class="bg-transparent card border-0 mx-3 px-1">
+              <h3>Currently Reading</h3>
+              <div class="d-flex align-items-center"></div>
+              <h3>To Be Read</h3>
+              <div class="d-flex align-items-center mb-2"><a class="mr-3" href="books/add">Add to the list</a><a href="/books/rand" id="pick-random">🎲 Select Next Read</a></div>
+              <div class="d-flex align-items-center"></div>
+              <h3>Completed</h3>
+            </div>
+          </div>
         </div>
-        <h3>To Be Read</h3>
-        <div class="d-flex align-items-center">
-          <ul id="book-list"></ul>
-        </div>
-        <h3>Completed</h3>
-        <ul id="completed-list"></ul>
       </div>
     </div>
-  </body>
+  </div>
+  <div class="text-center mt-4"><a href="/">🦶 Books Home</a></div>
+  <script>
+    function markAsRead() {
+      const currentBookId = event.target.id;
+      fetch(`/api/books/${currentBookId}`, {
+        method: 'PATCH',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({ readOn: new Date().toISOString(), currentlyReading: false })
+      })
+      .then((response) => {
+        if (!response.ok) {
+          throw new Error(response.message);
+        } else {
+          window.location.reload();
+        }
+      })
+    }
+    
+    function selectToRead() {
+      const selectedBookId = event.target.id;
+      fetch(`/api/books/${selectedBookId}`, {
+        method: 'PATCH',
+        headers: {
+          'Content-Type': 'application/json'
+        },
+        body: JSON.stringify({ currentlyReading: true })
+      })
+      .then((response) => {
+        if (!response.ok) {
+          throw new Error(response.message);
+        } else {
+          window.location.reload();
+        }
+      })
+    }
+  </script>
 </html>
