+++ views/random-book.html
@@ -1,0 +1,44 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <title>Random Book</title>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>🦶 Book List</title>
+    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous">
+  </head>
+  <body>
+    <div class="container mt-5">
+      <div class="row">
+        <div class="col-md-8 offset-md-2">
+          <h1 class="text-center mb-4">Random Book</h1>
+          <div class="card">
+            <div class="card-body">
+              <p class="text-muted">No book available.</p>
+            </div>
+          </div>
+          <div class="text-center mt-4"><a class="mr-3" href="/">🦶 Books Home</a><a href="/books">🦶 Books List</a></div>
+        </div>
+      </div>
+    </div>
+    <script>
+      function selectBook() {
+        const selectedBookId = event.target.id;
+        fetch(`/api/books/${selectedBookId}`, {
+          method: 'PATCH',
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify({ currentlyReading: true })
+        })
+        .then((response) => {
+          if (!response.ok) {
+            throw new Error(response.message);
+          } else {
+            window.location.assign('/books');
+          }
+        })
+      }
+    </script>
+  </body>
+</html>
