+++ views/random-book.pug
@@ -1,0 +1,43 @@
+doctype html
+html
+  head
+    title Random Book
+    meta(charset="UTF-8")
+    meta(name="viewport" content="width=device-width, initial-scale=1.0")
+    title 🦶 Book List
+    link(rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/css/bootstrap.min.css" integrity="sha384-TX8t27EcRE3e/ihU7zmQxVncDAy5uIKz4rEkgIXeMed4M0jlfIDPvg6uqKI2xXr2" crossorigin="anonymous")
+  body
+    .container.mt-5
+      .row
+        .col-md-8.offset-md-2
+          h1.text-center.mb-4 Random Book
+          .card
+            .card-body
+              if book
+                h2.card-title #{book.title}
+                p.card-text by #{book.author}
+                button.btn.btn-primary(onclick='selectBook()' id=book._id) Pick it!
+                button.btn.btn-secondary.ml-3(onclick='window.location.reload()' id=book._id) Try Again
+              else
+                p.text-muted No book available.
+          .text-center.mt-4
+            a(href='/').mr-3 🦶 Books Home
+            a(href='/books') 🦶 Books List
+    script.
+      function selectBook() {
+        const selectedBookId = event.target.id;
+        fetch(`/api/books/${selectedBookId}`, {
+          method: 'PATCH',
+          headers: {
+            'Content-Type': 'application/json'
+          },
+          body: JSON.stringify({ currentlyReading: true })
+        })
+        .then((response) => {
+          if (!response.ok) {
+            throw new Error(response.message);
+          } else {
+            window.location.assign('/books');
+          }
+        })
+      }
